# src/main/resources/static/boat4you_ws_entities_crud.openapi.yaml
openapi: 3.0.3
info:
  title: Boat4You Web Service
  description: |-
    API definition for Boat4You Web Service. \
    Contains operations related to Users
  version: 0.0.1
servers:
  - url: https://localhost:8443
    description: Local deployment
...
paths:
...
  /catalogue/manufacturers:
    get:
      tags:
        - Catalogue
      summary: Fetch all manufacturers or search them by name
      description: Fetch all manufacturers or search them by name
      operationId: getManufacturers
      parameters:
        - in: query
          name: name
          schema:
            type: string
          required: false
          description: Name or part of name of the wanted manufacturers
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                description: Fetch all manufacturers result
                required:
                  - content
                  - page
                properties:
                  content:
                    description: Manufacturers list
                    type: array
                    items:
                      $ref: '#/components/schemas/ManufacturerDto'
                  page:
                    $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/InternalPageModel'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/ErrorSchema'
        '500':
          description: Internal server error
  /catalogue/models:
    get:
      tags:
        - Catalogue
      summary: Fetch all models by manufacturerId
      description: Fetch all models by manufacturerId
      operationId: getModels
      parameters:
        - in: query
          name: manufacturerId
          schema:
            type: integer
            default: 0
            format: int64
          required: true
        - in: query
          name: name
          schema:
            type: string
          required: false
          description: Name or part of name of the wanted model
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                description: Fetch all models result
                required:
                  - content
                  - page
                properties:
                  content:
                    description: Models list
                    type: array
                    items:
                      $ref: '#/components/schemas/ModelDto'
                  page:
                    $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/InternalPageModel'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/ErrorSchema'
        '500':
          description: Internal server error
  /admin/agencies:
    get:
      tags:
        - Agencies
      summary: Fetch all agencies or search them by name
      description: This can only be done by the logged in user. Requires "SYSTEM_ADMIN" role.
      operationId: getAllAgencies
      parameters:
        - in: query
          name: name
          schema:
            type: string
          required: false
          example: SomeAgencyName
          description: Name or part of name of the wanted agency
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                description: Fetch all agencies or search by name result
                required:
                  - content
                  - page
                properties:
                  content:
                    description: Agencies list
                    type: array
                    items:
                      $ref: '#/components/schemas/AgencyDto'
                  page:
                    $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/InternalPageModel'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/ErrorSchema'
        '403':
          description: Not authorized
        '500':
          description: Internal server error
  /admin/agencies/{id}:
    get:
      tags:
        - Agencies
      summary: Get agency by agency id
      description: This can only be done by the logged in user. Requires "SYSTEM_ADMIN" role.
      operationId: getAgencyById
      parameters:
        - name: id
          in: path
          description: Id of the agency to be fetched
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgencyDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/ErrorSchema'
        '403':
          description: Not authorized
        '404':
          description: Agency not found
        '500':
          description: Internal server error
    put:
      tags:
        - Agencies
      summary: Update agency found by provided agency id
      description: This can only be done by the logged in user. Requires "SYSTEM_ADMIN" role.
      operationId: updateAgency
      parameters:
        - name: id
          in: path
          description: Id of the agency to be updated
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update an existing agency
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgencyDto'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgencyDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/ErrorSchema'
        '500':
          description: Internal server error
  /admin/agencies/{id}/activate:
    patch:
      tags:
        - Agencies
      summary: Toggles active property of wanted agency to true
      description: This can only be done by the logged in user. Requires "SYSTEM_ADMIN" role.
      operationId: activate
      parameters:
        - name: id
          in: path
          description: Id of the agency to toggle
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgencyDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/ErrorSchema'
        '403':
          description: Not authorized
        '500':
          description: Internal server error
  /admin/agencies/{id}/deactivate:
    patch:
      tags:
        - Agencies
      summary: Toggles active property of wanted agency to false
      description: This can only be done by the logged in user. Requires "SYSTEM_ADMIN" role.
      operationId: deactivate
      parameters:
        - name: id
          in: path
          description: Id of the agency to toggle
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgencyDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/ErrorSchema'
        '403':
          description: Not authorized
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ManufacturerDto:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    ModelDto:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
...
    AgencyDto:
      type: object
      required:
        - name
        - active
        - skipExternalSystem
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: Najbolja agencija
        address:
          type: string
          example: Domovinskog rata 5
        city:
          type: string
          example: Split
        country:
          type: string
          example: Croatia
        zip:
          type: string
          example: 21000
        vatCode:
          type: string
          example: B59915249
        web:
          type: string
          example: www.sealand.it
        email:
          type: string
          example: booking@caymangroupcharter.it
        phone:
          type: string
          example: +39 3936101443
        mobile:
          type: string
          example: +39 3936101443
        iban:
          type: string
          example: 12347890
        active:
          type: string
          example: 12347890
        discount:
          type: integer
          example: 10
        director:
          type: string
          example: Konstantina Chachali
        skipExternalSystem:
          type: boolean
          example: true
...
security:
  - bearerAuth: []